feat: setting to force oauth2/oidc login
