+++ apps/nestjs-backend/src/features/auth/local-auth/local-auth.module.ts
@@ -1,12 +1,15 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { UserModule } from '../../user/user.module';
-import { AuthModule } from '../auth.module';
+import { SessionStoreService } from '../session/session-store.service';
+import { SessionModule } from '../session/session.module';
 import { LocalStrategy } from '../strategies/local.strategy';
 import { LocalAuthController } from './local-auth.controller';
+import { LocalAuthService } from './local-auth.service';
 
 @Module({
-  imports: [UserModule, forwardRef(() => AuthModule)],
-  providers: [LocalStrategy],
+  imports: [UserModule, SessionModule],
+  providers: [LocalStrategy, LocalAuthService, SessionStoreService],
   controllers: [LocalAuthController],
+  exports: [LocalAuthService],
 })
 export class LocalAuthModule {}
