+++ apps/nestjs-backend/src/features/auth/session/session.module.ts
@@ -4,10 +4,12 @@ import passport from 'passport';
 import { SessionHandleModule } from './session-handle.module';
 import { SessionHandleService } from './session-handle.service';
 import { SessionStoreService } from './session-store.service';
+import { SessionService } from './session.service';
 
 @Module({
   imports: [SessionHandleModule],
-  providers: [SessionStoreService],
+  providers: [SessionService, SessionStoreService],
+  exports: [SessionService],
 })
 export class SessionModule implements NestModule {
   constructor(private readonly sessionHandleService: SessionHandleService) {}
