+++ apps/nestjs-backend/src/features/auth/strategies/local.strategy.ts
@@ -2,14 +2,14 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
 import { UserService } from '../../user/user.service';
-import { AuthService } from '../auth.service';
+import { LocalAuthService } from '../local-auth/local-auth.service';
 import { pickUserMe } from '../utils';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
   constructor(
     private readonly userService: UserService,
-    private readonly authService: AuthService
+    private readonly authService: LocalAuthService
   ) {
     super({
       usernameField: 'email',
